Updates after getting clarification regarding WebLinks
Motivation:
Got a reply regarding the question about the Weblink headers and the
location header. The response [1] was that both the location header and
two link headers should be returned in the response to a registration
request. The link header of type 'push:monitor' url will be the same as
the URL in the location header.

[1] http://www.ietf.org/mail-archive/web/webpush/current/msg00106.html

Modifications:
Updated the README.md clarify this and also updated the server and
client to handle this.

Will need to refactor the WebLink types into an enum but not quite sure
what the best place if for them so that will be a separate commit.