Adding an example that works in Firefox nightly.
Note that there is currently no support for the WebPush API and this is
an attempt to get things working anyway. Currently I need to figure out
how HTTP/2 server push will work with browsers. I think that for push to
be used the WebPush API support would have to leave in the native
browser implementation and have access to the push stream that is
established by the server when monitor is called.