Adding response headers for monitor request as per spec.
Motivation:
The specification states that a monitor request also return the channel
weblink and expiration information.
This is mentioned in section 4 [1] which is about registering:
"The push server SHOULD also provide the "channel" link and expiration
   information in response to requests to the "monitor" resource."

This is not mentioned in monitoring section [2] though.

[1] https://tools.ietf.org/html/draft-thomson-webpush-http2-01#section-4
[2] https://tools.ietf.org/html/draft-thomson-webpush-http2-01#section-6

Result:
A monitor request will now return the channel weblink header and also
'cache-control' header