Changed names to be inline with draft-02
Motivation:
The specification has been updated to draft-02 [1] and it contains a
number of name changes. For example, channels are not mentioned but
instead subscription is used. In a similar manner the resource is not
'subscribe' instread of 'channel'.

[1] http://tools.ietf.org/html/draft-thomson-webpush-http2-02

Modifications:
Removed the refs to channel and using subscribe and subscriptions in
it's replace.
Also added the correct WebLink types.