Re-adding the aggregate command.
Motivation:
I removed the aggregate command as it used classes from server-netty
which is written in Java 1.8. Until node-webpush-server support ALPN we
have to be able to use NPN with the client which for Java means that we
have to stick to Java 1.7. But we still want to be able to use the
aggregate function with our server so I'm adding it back and duplicating
some code from server-netty with some minor changes removing the Java
1.8 specifics.

Modifications:
Added back the aggregate command.

Result:
It is now possible to used the aggregate comment with WebPush server
that support it.