Adding the copying of ALPN and NPN libraries.
Motivation:
The version of ALPN/NPN are detected depending on the version of Java in
use. The correct versions were being detected but the dependencies
themselves were never downloaded into the local maven repository.

Modifications:
Added a copy-dependency task to be executed so that the libraries are
resolved and downloaded.

Result:
Is not possible to run the WebPush Console and the server.