Trying to improve error handling.
Motivation:
Currently, if the TLS handshake fails during SNI record/process it still
looks like the client is connected. There will only be an error when the
following command is executed.

Modifications:
This commit adds a message to the console if the channel is disconnected
and includes the channelId so that the user can see which channel was
disconnected. Also added the channelId to the output when connecting for
ref.

Result:
Slightly more information about the state of a connection.