Adding zip distributions for WebPush Console
Motivation:
Currently the process of running the WebPush Console is pretty
complicated. This commit aim to simplify the work needed to get the
console running.

Modifications:
We now provide distributions of the console specific for
different java versions. These distributions will include the correct
NPN or ALPN TLS extension libs and also a shell script to start the
server. The distributions also contain a readme file the shows the
command to start the console for evironments that don't support shell
scripts.

Result:
To start a console for a specific version of Java it is only a matter
unzipping and running the shell script:

$ unzip -d java7_25 webpush-console-7_0_25.zip
$ cd java7_25
$ ./console.sh