Change Netty version to 4.1.0.beta4-SNAPSHOT.
Motivation:
We are currently using version 5.0.0.Alpha2-SNAPSHOT. But to release we
need a non-snapshot version. There will be a 4.1.0.beta4 release shortly
and we should update our code to make sure it work with that version.
This will make it easy for us to change to the non-snapshot version once
it is available.

Modifications:
Updated the version in the pom.xml and make modifications required for
the difference in Netty code base between 5.0.0 and 4.1.

Result:
Everything works as before and we'll be able to switch to 4.1.0.beta4
when it is available.

JIRA:
https://issues.jboss.org/browse/AGSYNC-46