Adding missing NPN distributions and adding docs
Motivation:
Currenlty not all Java 1.7 version and NPN combinations are available.
Also the documentation is not clear about this requirement which is has
become evident after pull request reviews.

Modifications:
Added the missing zip distributions and update console/readme with the
combinations spans and links to NPNs docs for further information.

Result:
Most Java 1.7 minor version are supported. There are a few older ones
that are not covered. But more importantly the lastest version are.