Allow WebPushFrameListener to take encoder as constructor parameter
Motivation:
Currently the WebPushFrameListener only takes a WebPushServer instance
as a parameter and has a method named encoder to set the
Http2ConnectionEncoder to be used. The was a left over from an older
version of Netty where this was the way to set the encoder.

Modifications:
Added the encoder as a parameter.

Result:
All test pass.